make the public API clear
