+++ crates/fluvio-smartengine/src/engine/memory.rs
