+++ crates/fluvio-smartengine/src/engine/transforms/map.rs
@@ -8,41 +8,42 @@ mod test {
         Record,
     };
 
-    use crate::{
+    use crate::engine::{
         SmartEngine, SmartModuleChainBuilder, SmartModuleConfig, metrics::SmartModuleChainMetrics,
-        transforms::simple_transform::FILTER_MAP_FN_NAME,
+        transforms::simple_transform::MAP_FN_NAME,
     };
+    use crate::engine::fixture::read_wasm_module;
 
-    const SM_FILTER_MAP: &str = "fluvio_smartmodule_filter_map";
-
-    use crate::fixture::read_wasm_module;
+    const SM_MAP: &str = "fluvio_smartmodule_map";
 
     #[ignore]
     #[test]
-    fn test_filter_map() {
+    fn test_map() {
         let engine = SmartEngine::new();
         let mut chain_builder = SmartModuleChainBuilder::default();
 
         chain_builder.add_smart_module(
             SmartModuleConfig::builder().build().unwrap(),
-            read_wasm_module(SM_FILTER_MAP),
+            read_wasm_module(SM_MAP),
         );
 
         let mut chain = chain_builder
             .initialize(&engine)
-            .expect("failed to build chain");
+            .expect("failed to build chain")
+            .inner;
 
         assert_eq!(
             chain.instances().first().expect("first").transform().name(),
-            FILTER_MAP_FN_NAME
+            MAP_FN_NAME
         );
 
         let metrics = SmartModuleChainMetrics::default();
-        let input = vec![Record::new("10"), Record::new("11")];
+        let input = vec![Record::new("apple"), Record::new("fruit")];
         let output = chain
             .process(SmartModuleInput::try_from(input).expect("input"), &metrics)
             .expect("process");
-        assert_eq!(output.successes.len(), 1); // one record passed
-        assert_eq!(output.successes[0].value.as_ref(), b"5");
+        assert_eq!(output.successes.len(), 2); // one record passed
+        assert_eq!(output.successes[0].value.as_ref(), b"APPLE");
+        assert_eq!(output.successes[1].value.as_ref(), b"FRUIT");
     }
 }
